refactor(auth): refactor ui components
