+++ src/app/(auth)/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginUI } from '@/components/lms/auth/components/login.ui';
+import { LoginUI } from '@/components/lms/auth/components/login';
 
 export default function Page() {
   return <LoginUI />;
