+++ src/app/(auth)/register/page.tsx
@@ -1,4 +1,4 @@
-import { RegisterUI } from '@/components/lms/auth/components/register.ui';
+import { RegisterUI } from '@/components/lms/auth/components/register';
 
 export default function Page() {
   return <RegisterUI />;
